Removed useless route Added protection in db migration for the case no backing project is configured Fixes #20

+++ db/migrate/011_ts_remove_issues.rb
@@ -3,8 +3,8 @@
 
 class TsRemoveIssues < ActiveRecord::Migration
   def up
-    ts_prj = Project.find(Setting.plugin_redmine_app_timesheets['project'].to_i)
-    ts_prj.issues.destroy_all
+    ts_prj = Project.find(Setting.plugin_redmine_app_timesheets['project'].to_i) rescue nil
+    ts_prj.issues.destroy_all unless ts_prj.nil?
   end
 
   def down
